+++ facebook/__init__.py
@@ -168,25 +168,6 @@ class GraphAPI(object):
             post_args=data,
             method="POST")
 
-    def put_wall_post(self, message, attachment={}, profile_id="me"):
-        """Writes a wall post to the given profile's wall.
-
-        We default to writing to the authenticated user's wall if no
-        profile_id is specified.
-
-        attachment adds a structured attachment to the status message
-        being posted to the Wall. It should be a dictionary of the form:
-
-            {"name": "Link name"
-             "link": "https://www.example.com/",
-             "caption": "{*actor*} posted a new review",
-             "description": "This is a longer description of the attachment",
-             "picture": "https://www.example.com/thumbnail.jpg"}
-
-        """
-        return self.put_object(profile_id, "feed", message=message,
-                               **attachment)
-
     def put_comment(self, object_id, message):
         """Writes the given comment on the given post."""
         return self.put_object(object_id, "comments", message=message)