+++ docs/changes.rst
@@ -13,6 +13,7 @@ Version 3.0.0 (unreleased)
  - Add new `get_all_connections` method to make pagination easier (#337).
  - Add new `get_permissions` method to retrieve permissions that a user has
    granted an application (#264, #342).
+ - Remove `put_wall_post` method. Use `put_object` instead.
 
 Version 2.0.0 (2016-08-08)
 ==============================